Clock goes back to the speaking-mode SL #4079
This PATCH is just a workaround to the real problem:

     Clock is using gst 0.11 but we have installed the 'espeak' plugin
     for Gst 1.0 because we need that version for the shell
     itself. So, Clock does not find the proper plugin and tries to
     use the 'espeak' command directily but it fails as well because a
     'portaudio' lib issue.

This PATCH makes 'espeak' to create a WAV file in a temporary file and
then plays it with the command 'playwave'. It's the same idea used in
Speak, with the difference than Speak uses gstreamer to play the wav
file.

This workaround will not be necessary when Clock uses Gtk3 and Gst
1.0.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>