Draw the simple background only when the widget resizes - SL #1959
The static elements of the simple clock can be drawn each time the
widget resizes, instead of for each tick of the clock.  This are the
white circle background, the ticks for hour and minutes, and the
numbers.  This patch adds a helper surface to do it, the same as the
nice clock is doing.  The activity gets a performance improvement
doing this.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>