Emit time after the hands of the clock were grabbed - SL #1959
This provokes an update in the displayed time in full letters and also
speaks the time if the speech time option is toggled.

One decision had to be made: which time should it output? AM or PM?
Because from the analog clock both can be tell.  I went for the PM
because looks more intuitive for children.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>