Port to Python 3 - remove old code
Port to Python 3 is on a baseline of Toolkit 0.114, which certainly has
sugar3.graphics.progresstoolbutton, so remove the import failure
detection, and the activity local copy of the code.

Toolkit 0.114 also certainly has sugar3.speech, so remove the direct use
of espeak.

Port to Python 3 will not be deployed to OLPC XO laptops any time soon,
so remove the check for laptop model and the user interface mode
required for limiting number of GStreamer playback pipelines.