Fix ticking clock clicks
On OLPC XO-4 laptop, opening the audio device causes a click, and this
makes the ticking clock noisy.

Existing logic was to play a 200 ms sound file every second, wait for
end of stream, rewind, and play again.  This closed and opened the audio
device.

Now, rewind and play a 1200 ms sound file and pause it after 450 ms.
End of stream is never reached.  This keeps the audio device open.